\name{icar-package}
\alias{ICAR-package}
\alias{icar}
\alias{ICAR}
\docType{package}
\title{
A package for the generation and analysis of cognitive ability items.
}
\description{
The ICAR package has been developed to include functions for the generation of cognitive ability item types.  The ICAR package is intended to facilitate the distribution of cognitive ability measures which have been described in several locations as the International Cognitive Ability Resource.  While some of the functions in this package are also useful for data entry, most will be used to generate output that is compatible with LaTeX software.  The resulting LaTeX code can then be used to generate large item sets for use across different platforms.

To see a list of all functions, use:
help(package="icar").

To cite this package, use:
citation(package = "icar", lib.loc = NULL)

For more information about item types or their administration, please contact one of the authors.
}
\details{
\tabular{ll}{
Package: \tab ICAR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-01-15\cr
License: \tab GPL (>=3)\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Jason A. French, David M. Condon & William Revelle\cr
Department of Psychology\cr
Northwestern University\cr
Evanston, Illinois\cr
\tabular{ll}{
Maintainer: \tab Jason A. French <frenchja@u.northwestern.edu>\cr
\tab \url{https://github.com/frenchja/icar}\cr
}
}
\references{
Revelle, W. (in preparation). An Introduction to Psychometric Theory with applications in R. Springer. Retrieved from \url{http://personality-project.org/r/book/}

Condon, D.M. & Revelle, W.  (in manuscript).  The International Cognitive Ability Resource: Development and initial validation of a public-domain measure.
}
\keyword{ psychometrics }
\keyword{ ability }
\keyword{ generation }
\seealso{

\code{\link[psych-package]{psych}}
}
\examples{

}
